% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-africa-data.R
\name{create_africa_data}
\alias{create_africa_data}
\title{Create malaria prevalence data}
\usage{
create_africa_data()
}
\value{
a dataset
}
\description{
Create malaria prevalence data
}
\examples{

create_africa_data()
}
